requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Hydra configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging and visualization
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ML utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Progress bars and CLI
tqdm>=4.65.0
rich>=13.0.0

# Development tools
pytest>=7.3.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.3.0

# Environment
python-dotenv>=1.0.0
